+++ frontend/components/landing/FeatureShowcase.tsx
@@ -180,12 +180,6 @@ export default function FeatureShowcase() {
 
   return (
     <section className="relative pt-24 pb-12 lg:pt-32 lg:pb-16">
-      {/* Background decoration */}
-      <div className="absolute inset-0 -z-10 overflow-hidden">
-        <div className="absolute top-1/4 left-0 w-96 h-96 bg-blue-500/5 rounded-full blur-3xl" />
-        <div className="absolute bottom-1/4 right-0 w-96 h-96 bg-violet-500/5 rounded-full blur-3xl" />
-      </div>
-
       <div className="max-w-7xl mx-auto px-6 lg:px-8">
         {/* Section header */}
         <motion.div
